[validation] Added a validator that checks the annotations
Also:

- Fixed name for inline response. This was a, potentially, changing number based name. Now it is the operation id + Response.

- Generate warning when the tags are not set and the Configuration config.tagsMustBeSet is true.

+++ biz.aQute.openapi.generator/src/aQute/openapi/annotations/ValidatorNumber.java
@@ -18,6 +18,6 @@ public @interface ValidatorNumber {
 
 	boolean exclusiveMinimum() default false;
 
-	double multipleOf() default 1;
+	long multipleOf() default 1;
 
 }
